This plugin is responsible for not calling <l:task>
While working on 1.580.1 I noticed that the new version of the credentials plugin was breaking a test by
rendering an action "Credentials" that has a broken link.

A research seems to indicate that this action isn't meant to be shown, as the getIconFile() method
was returning null, and <l:task> never seems to have accepted @icon as null (Action.getIconFile() says
it can return null to hide itself, but that logic was handled in <l:tasks>.)

So AFAICT, this image rendering has already been broken.