[JENKINS-66514] BaseStandardCredentials ID validation error
See details in JENKINS-66514

The fix is to remove the unnecesary explicit `checkUrl` so the legacy
validation method is nt used. The check URL will be calculated
automatically and the "new" (years old) mechanism kicks in.

The test is checking that the validation error shows up when needed (and
only when needed).