[JENKINS-68616] Enable field validation checks
Regression in 1105 release.  The `div` that was added to surround the
`select` seems to prevent calls to the `attrs.checkMethod`.

When using a GitHub branch source, an empty credentials field no longer
displays the warning that a credential is recommended.  Other uses of
credentials seem to have the same behavior.  Validation checks are no
longer called.

When the `div` is included in the page, the credential validation callback
is not invoked. When the `div` is removed from the page, the credential
validation callback is invoked.

It is unclear to me why the addition of a `div` would change the
validation behavior.

Review the differences with white space changes suppressed so that it
is clear the difference is only in the deletion of the `div` tags that
wrapped the `select`.

Revert one HTML `div` from "Minor improvements"

This reverts part of commit 6dee9b8602f4ba355209f46ddd642d6daca3b448.