Remove FileOnMasterKeySource and SECURITY-1322 migration
It has been 5 years since the security fix was introduced and the
migration code occureed.

Whilst an admin could have still (ab)used the CLI/REST to set a
FileOnMaster they could not do this from the UI creating a disparity.

This change removes the migration and the ability to use this via
CLI/REST